requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
distribute==0.6.27
python-customerio==0.1.1
requests==1.1.0
wsgiref==0.1.2
